\name{breakpointError-package}
\alias{breakpointError-package}
\docType{package}
\title{Breakpoint detection error}
\description{Functions for calculating the breakpoint detection error,
which quantifies the number of false positive and false
negative breakpoint detections when the latent signal
is known.}
\details{
\tabular{ll}{Package: \tab breakpointError\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Breakpoint detection error\cr
Suggests: \tab cghseg, ggplot2, directlabels, quadprog, tikzDevice, reshape2\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


